ex01 done

+++ Module_02/ex01/Fixed.cpp
@@ -16,7 +16,7 @@ Fixed::Fixed( const int num ) {
 }
 
 Fixed::Fixed( const float num ) {
-    this->setRawBits( round( num * ( 1 << _bitsNum ) ));
+    this->setRawBits( ( int )round( num * ( 1 << _bitsNum ) ));
     std::cout << "Float constructor called" << std::endl;
 }
